uploader/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
loguru==0.7.2
python-multipart